Moved translator-related files to subdirectory
